refactor: Move component types to core directory
