+++ src/types/core/component.ts
@@ -1,10 +1,12 @@
+import { EventObject } from './event';
+
 export interface CurrentDoms {
   [key: string]: string;
 }
 
 export interface UseCurrentDom {
-  getCurrentDom(): string;
-  setCurrentDom(newCurrentDom: string): string;
+  getCurrentDom: () => string;
+  setCurrentDom: (newCurrentDom: string) => string;
 }
 
 export interface ComponentProps {
@@ -13,8 +15,9 @@ export interface ComponentProps {
 
 export type Component = {
   (props?: ComponentProps): {
-    render(): string;
+    render: () => string;
     getCurrentDom: UseCurrentDom['getCurrentDom'];
     setCurrentDom: UseCurrentDom['setCurrentDom'];
+    setEvent?: (args?: EventObject) => void;
   };
 };
